@@ -3,6 +3,8 @@ package api
 import (
 	"reflect"
 	"testing"
+
+	"github.com/pelletier/go-toml/v2"
 )
 
 func toInlineDocumentMust(m map[string]interface{}) InlineDocument {
@@ -11,6 +13,20 @@ func toInlineDocumentMust(m map[string]interface{}) InlineDocument {
 		panic(err)
 	}
 	return d
+}
+
+// pass the toml through serialization and deserialization to get a normalized
+// payload for tests that has deterministic ordering and formatting
+func tomlNormalize(t string) string {
+	var m map[string]interface{}
+	if err := toml.Unmarshal([]byte(t), &m); err != nil {
+		panic(err)
+	}
+	s, err := toml.Marshal(m)
+	if err != nil {
+		panic(err)
+	}
+	return string(s)
 }
 
 func TestMerge(t *testing.T) {
@@ -45,6 +61,26 @@ func TestMerge(t *testing.T) {
 			},
 		},
 		{
+			name: "merge with deeply nested toml object",
+			baseSpec: NodeConfigSpec{
+				Containerd: ContainerdOptions{
+					Config: "[a.b.c.d]\nf = 0",
+				},
+			},
+			patchSpec: NodeConfigSpec{
+				Containerd: ContainerdOptions{
+					Config: "[a.b.c.d]\ne = 0",
+				},
+			},
+			// This test is primarily for clarity on what happens during the
+			// expansion of nested toml objects.
+			expectedSpec: NodeConfigSpec{
+				Containerd: ContainerdOptions{
+					Config: "[a]\n[a.b]\n[a.b.c]\n[a.b.c.d]\ne = 0\nf = 0\n",
+				},
+			},
+		},
+		{
 			name: "customer overrides orchestrator defaults",
 			baseSpec: NodeConfigSpec{
 				Cluster: ClusterDetails{
@@ -66,7 +102,33 @@ func TestMerge(t *testing.T) {
 					},
 				},
 				Containerd: ContainerdOptions{
-					Config: "base",
+					Config: tomlNormalize(`
+version = 2
+root = "/var/lib/containerd"
+state = "/run/containerd"
+
+[grpc]
+address = "/run/containerd/containerd.sock"
+
+[plugins."io.containerd.grpc.v1.cri".containerd]
+default_runtime_name = "runc"
+discard_unpacked_layers = true
+
+[plugins."io.containerd.grpc.v1.cri"]
+sandbox_image = "{{.SandboxImage}}"
+
+[plugins."io.containerd.grpc.v1.cri".registry]
+config_path = "/etc/containerd/certs.d:/etc/docker/certs.d"
+
+[plugins."io.containerd.grpc.v1.cri".containerd.runtimes.runc]
+runtime_type = "io.containerd.runc.v2"
+
+[plugins."io.containerd.grpc.v1.cri".containerd.runtimes.runc.options]
+SystemdCgroup = true
+
+[plugins."io.containerd.grpc.v1.cri".cni]
+bin_dir = "/opt/cni/bin"
+conf_dir = "/etc/cni/net.d"`),
 				},
 			},
 			patchSpec: NodeConfigSpec{
@@ -82,7 +144,13 @@ func TestMerge(t *testing.T) {
 					},
 				},
 				Containerd: ContainerdOptions{
-					Config: "patch",
+					Config: tomlNormalize(`
+version = 2
+[grpc]
+address = "/run/containerd/containerd.sock.2"
+
+[plugins."io.containerd.grpc.v1.cri".containerd]
+discard_unpacked_layers = false`),
 				},
 			},
 			expectedSpec: NodeConfigSpec{
@@ -107,7 +175,33 @@ func TestMerge(t *testing.T) {
 					},
 				},
 				Containerd: ContainerdOptions{
-					Config: "patch",
+					Config: tomlNormalize(`
+version = 2
+root = "/var/lib/containerd"
+state = "/run/containerd"
+
+[grpc]
+address = "/run/containerd/containerd.sock.2"
+
+[plugins."io.containerd.grpc.v1.cri".containerd]
+default_runtime_name = "runc"
+discard_unpacked_layers = false
+
+[plugins."io.containerd.grpc.v1.cri"]
+sandbox_image = "{{.SandboxImage}}"
+
+[plugins."io.containerd.grpc.v1.cri".registry]
+config_path = "/etc/containerd/certs.d:/etc/docker/certs.d"
+
+[plugins."io.containerd.grpc.v1.cri".containerd.runtimes.runc]
+runtime_type = "io.containerd.runc.v2"
+
+[plugins."io.containerd.grpc.v1.cri".containerd.runtimes.runc.options]
+SystemdCgroup = true
+
+[plugins."io.containerd.grpc.v1.cri".cni]
+bin_dir = "/opt/cni/bin"
+conf_dir = "/etc/cni/net.d"`),
 				},
 			},
 		},
