@@ -1,5 +1,4 @@
 //go:build !ignore_autogenerated
-// +build !ignore_autogenerated
 
 // Code generated by controller-gen. DO NOT EDIT.
 
@@ -37,6 +36,13 @@ func (in *ClusterDetails) DeepCopy() *ClusterDetails {
 // DeepCopyInto is an autogenerated deepcopy function, copying the receiver, writing into out. in must be non-nil.
 func (in *ContainerdOptions) DeepCopyInto(out *ContainerdOptions) {
 	*out = *in
+	if in.BaseRuntimeSpec != nil {
+		in, out := &in.BaseRuntimeSpec, &out.BaseRuntimeSpec
+		*out = make(InlineDocument, len(*in))
+		for key, val := range *in {
+			(*out)[key] = *val.DeepCopy()
+		}
+	}
 }
 
 // DeepCopy is an autogenerated deepcopy function, copying the receiver, creating a new ContainerdOptions.
@@ -221,7 +227,7 @@ func (in *NodeConfigList) DeepCopyObject() runtime.Object {
 func (in *NodeConfigSpec) DeepCopyInto(out *NodeConfigSpec) {
 	*out = *in
 	in.Cluster.DeepCopyInto(&out.Cluster)
-	out.Containerd = in.Containerd
+	in.Containerd.DeepCopyInto(&out.Containerd)
 	out.Instance = in.Instance
 	in.Kubelet.DeepCopyInto(&out.Kubelet)
 	if in.FeatureGates != nil {
